requirements.txt
aniso8601==7.0.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-rest-auth==0.9.5
django-summernote==0.8.11.4
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==20.0.0
idna==2.8
oauthlib==3.1.0
Pillow==6.2.0
promise==2.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
